# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Teston HR Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentListResponse"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDepartmentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentResponse"
  /departments/{id}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get a department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDepartmentRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepartmentResponse"
...
  /leaves:
    get:
      operationId: listLeaves
      tags: [leaves]
      summary: List leave applications
      parameters:
        - in: query
          name: userId
          schema:
            type: number
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: leaveType
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
        - in: query
          name: endDate
          schema:
            type: string
      responses:
        "200":
          description: Leave applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveListResponse"
    post:
      operationId: createLeave
      tags: [leaves]
      summary: Submit a leave application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeaveRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leaves/{id}:
    get:
      operationId: getLeave
      tags: [leaves]
      summary: Get a leave application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Leave application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateLeave
      tags: [leaves]
      summary: Review/update a leave application
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLeaveRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveResponse"
  /leaves/balance/{userId}:
    get:
      operationId: getLeaveBalance
      tags: [leaves]
      summary: Get leave balance for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: number
        - in: query
          name: year
          schema:
            type: number
      responses:
        "200":
          description: Leave balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveBalanceResponse"
  /medical-certificates:
    get:
      operationId: listMedicalCertificates
      tags: [medical-certificates]
      summary: List medical certificates
      parameters:
        - in: query
          name: userId
          schema:
            type: number
        - in: query
          name: verificationStatus
          schema:
            type: string
      responses:
        "200":
          description: Medical certificates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalCertificateListResponse"
    post:
      operationId: createMedicalCertificate
      tags: [medical-certificates]
      summary: Upload a medical certificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMedicalCertificateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalCertificateResponse"
  /medical-certificates/{id}:
    get:
      operationId: getMedicalCertificate
      tags: [medical-certificates]
      summary: Get a medical certificate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Medical certificate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalCertificateResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateMedicalCertificate
      tags: [medical-certificates]
      summary: Update a medical certificate
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMedicalCertificateRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicalCertificateResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    DepartmentResponse:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    DepartmentListResponse:
      type: object
      properties:
        departments:
          type: array
          items:
            $ref: "#/components/schemas/DepartmentResponse"
        total:
          type: number
      required:
        - departments
        - total
    CreateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
      required:
        - name
    UpdateDepartmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...
    LeaveResponse:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        totalDays:
          type: string
        reason:
          type: string
        status:
          type: string
        reviewedBy:
          type: number
          nullable: true
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        reviewNotes:
          type: string
          nullable: true
        aiClassification:
          type: string
          nullable: true
        aiConfidence:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - leaveType
        - startDate
        - endDate
        - totalDays
        - reason
        - status
        - createdAt
        - updatedAt
    LeaveListResponse:
      type: object
      properties:
        leaves:
          type: array
          items:
            $ref: "#/components/schemas/LeaveResponse"
        total:
          type: number
      required:
        - leaves
        - total
    CreateLeaveRequest:
      type: object
      properties:
        userId:
          type: number
        leaveType:
          type: string
          enum: [annual, medical, emergency, unpaid, other]
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
      required:
        - userId
        - leaveType
        - startDate
        - endDate
        - reason
    UpdateLeaveRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, cancelled]
        reviewNotes:
          type: string
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
    LeaveBalanceResponse:
      type: object
      properties:
        userId:
          type: number
        year:
          type: number
        annualLeaveUsed:
          type: number
        annualLeaveTotal:
          type: number
        annualLeaveRemaining:
          type: number
        medicalLeaveUsed:
          type: number
        medicalLeaveTotal:
          type: number
        medicalLeaveRemaining:
          type: number
      required:
        - userId
        - year
        - annualLeaveUsed
        - annualLeaveTotal
        - annualLeaveRemaining
        - medicalLeaveUsed
        - medicalLeaveTotal
        - medicalLeaveRemaining
    MedicalCertificateResponse:
      type: object
      properties:
        id:
          type: number
        userId:
          type: number
        leaveApplicationId:
          type: number
          nullable: true
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        doctorName:
          type: string
          nullable: true
        clinicName:
          type: string
          nullable: true
        issueDate:
          type: string
          nullable: true
        mcStartDate:
          type: string
          nullable: true
        mcEndDate:
          type: string
          nullable: true
        verificationStatus:
          type: string
        verificationNotes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - fileName
        - fileUrl
        - fileType
        - verificationStatus
        - createdAt
        - updatedAt
    MedicalCertificateListResponse:
      type: object
      properties:
        certificates:
          type: array
          items:
            $ref: "#/components/schemas/MedicalCertificateResponse"
        total:
          type: number
      required:
        - certificates
        - total
    CreateMedicalCertificateRequest:
      type: object
      properties:
        userId:
          type: number
        leaveApplicationId:
          type: number
        fileName:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        doctorName:
          type: string
        clinicName:
          type: string
        issueDate:
          type: string
        mcStartDate:
          type: string
        mcEndDate:
          type: string
      required:
        - userId
        - fileName
        - fileUrl
        - fileType
    UpdateMedicalCertificateRequest:
      type: object
      properties:
        doctorName:
          type: string
        clinicName:
          type: string
        issueDate:
          type: string
        mcStartDate:
          type: string
        mcEndDate:
          type: string
        verificationStatus:
          type: string
          enum: [pending, verified, suspicious, unreadable]
        verificationNotes:
          type: string
        leaveApplicationId:
          type: number
...